requirements.txt
# Core dependencies
earthengine-api>=0.1.384
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.14.0
shapely>=2.0.0

# Database (for future integration)
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Logging
loguru>=0.7.0
